--- a/src/globby.nim
+++ b/src/globby.nim
@@ -31,15 +31,15 @@ proc add*[T](tree: GlobTree[T], path: string, data: T) =
     data: data
   ))
 
-proc globMatchOne(s, glob: string): bool =
+proc globMatchOne(path, glob: string, pathStart = 0, globStart = 0): bool =
   ## Match a single entry string to glob.
 
   proc error() =
     raise newException(GlobbyError, "Invalid glob: `" & glob & "`")
 
   var
-    i = 0
-    j = 0
+    i = pathStart
+    j = globStart
   while j < glob.len:
     if glob[j] == '?':
       discard
@@ -51,8 +51,8 @@ proc globMatchOne(s, glob: string): bool =
           inc j
         else:
           break
-      for k in i ..< s.len:
-        if globMatchOne(s[k .. ^1], glob[(j + 1) .. ^1]):
+      for k in i ..< path.len:
+        if globMatchOne(path, glob, k, j + 1):
           i = k - 1
           return true
       return false
@@ -62,14 +62,14 @@ proc globMatchOne(s, glob: string): bool =
       if j < glob.len and glob[j] == ']': error()
       if j + 3 < glob.len and glob[j + 1] == '-' and glob[j + 3] == ']':
         # Do [A-z] style match.
-        if s[i].ord < glob[j].ord or s[i].ord > glob[j + 2].ord:
+        if path[i].ord < glob[j].ord or path[i].ord > glob[j + 2].ord:
           return false
         j += 3
       else:
         # Do [ABC] style match.
         while true:
           if j >= glob.len: error()
-          elif glob[j] == s[i]:
+          elif glob[j] == path[i]:
             while glob[j] != ']':
               if j + 1 >= glob.len: error()
               inc j
@@ -78,18 +78,18 @@ proc globMatchOne(s, glob: string): bool =
           elif glob[j] == ']':
             return false
           inc j
-    elif i >= s.len:
+    elif i >= path.len:
       return false
-    elif glob[j] != s[i]:
+    elif glob[j] != path[i]:
       return false
     inc i
     inc j
 
-  if i == s.len and j == glob.len:
+  if i == path.len and j == glob.len:
     return true
 
 proc globMatch(
-  pathParts, globParts: seq[string], pathStart, globStart: int
+  pathParts, globParts: seq[string], pathStart = 0, globStart = 0
 ): bool =
   ## Match a seq string to a seq glob pattern.
   var
@@ -114,9 +114,6 @@ proc globMatch(
 
   if i == pathParts.len and j == globParts.len:
     return true
-
-proc globMatch(pathParts, globParts: seq[string]): bool =
-  globMatch(pathParts, globParts, 0, 0)
 
 proc globSimplify(globParts: seq[string]): seq[string] =
   ## Simplify backwards ".." and absolute "//".