--- a/tests/benchmark.nim
+++ b/tests/benchmark.nim
@@ -1,4 +1,4 @@
-import benchy, globby, random
+import benchy, globby, random, strutils
 
 randomize()
 
@@ -32,5 +32,51 @@ timeIt "findAll":
 timeIt "findAll **":
   var count: int
   for path in tree.findAll("**"):
+    inc count
+  doAssert count > 0
+
+timeIt "findAll **/..":
+  var path: string
+  while true:
+    path = paths[rand(0..paths.high)]
+    if path.contains('/'):
+      break
+
+  let glob = "**/" & path[path.find('/') + 1 ..< path.len]
+
+  var count: int
+  for path in tree.findAll(glob):
+    inc count
+  doAssert count > 0
+
+timeIt "findAll */..":
+  var path: string
+  while true:
+    path = paths[rand(0..paths.high)]
+    if path.contains('/'):
+      break
+
+  let glob = "*/" & path[path.find('/') + 1 ..< path.len]
+
+  var count: int
+  for path in tree.findAll(glob):
+    inc count
+  doAssert count > 0
+
+timeIt "findAll ../*/..":
+  var path: string
+  while true:
+    path = paths[rand(0..paths.high)]
+    if path.count('/') >= 2:
+      break
+
+  let
+    firstSlash = path.find('/')
+    secondSlash = firstSlash + 1 + path[firstSlash + 1 ..< path.len].find('/')
+
+  let glob = path[0 ..< firstSlash] & "/*/" & path[secondSlash + 1 ..< path.len]
+
+  var count: int
+  for path in tree.findAll(glob):
     inc count
   doAssert count > 0