before:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.081 ms      0.423 ms    ±0.259  x1000
findAll ** ......................... 0.063 ms      0.065 ms    ±0.002  x1000
findAll **/.. ...................... 2.168 ms      4.905 ms    ±2.128   x950
findAll */.. ....................... 0.151 ms      0.457 ms    ±0.239  x1000
findAll ../*/.. .................... 0.200 ms      0.518 ms    ±0.258  x1000
```

after:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.026 ms      0.028 ms    ±0.002  x1000
findAll ** ......................... 0.005 ms      0.006 ms    ±0.000  x1000
findAll **/.. ...................... 0.270 ms      0.280 ms    ±0.007  x1000
findAll */.. ....................... 0.025 ms      0.029 ms    ±0.004  x1000
findAll ../*/.. .................... 0.027 ms      0.044 ms    ±0.012  x1000
```